feat: Change to import sytnax for files and update resolvers/typeResolvers
