+++ src/resolvers/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+export { default as players } from './players';
