+++ src/server.js
@@ -1,40 +1,15 @@
-const { ApolloServer } = require('apollo-server');
-const fetch = require('node-fetch');
-const { importSchema } = require('graphql-import');
-const memoize = require('memoizee');
-
+import { ApolloServer } from 'apollo-server';
+import { importSchema } from 'graphql-import';
 // resolvers
-const typeResolvers = require('./resolvers/typeResolvers');
-const query = require('./resolvers/QueryResolver');
-// utils
-const sortPlayers = require('./utils/sortPlayers');
+import * as queryResolvers from './resolvers';
+import * as typeResolvers from './resolvers/typeResolvers';
 // schema
 const typeDefs = importSchema('./schemas/app.graphql');
 
 const resolvers = {
-  Query: {
-    async players(_, { per_page: perPage, page, sortOption, sortOrder, filter }) {
-      try {
-        const getData = () =>
-          fetch('https://fantasy.premierleague.com/drf/bootstrap-static').then(res => res.json());
-        const memoizedData = memoize(getData, { promise: true, maxAge: 1000 * 60 * 60 });
-        const data = await memoizedData();
-        const sortedData = sortOption
-          ? data.elements.sort(sortPlayers(sortOption, sortOrder))
-          : data.elements;
-        const updatedPage = page - 1;
-        return sortedData.slice(updatedPage * perPage, perPage * page);
-        // return data.elements
-      } catch (err) {
-        throw new Error(err);
-      }
-    }
-  },
-  ...typeResolvers
-};
-// TODO why doesnt this work?
-const resolvers2 = {
-  Query: query,
+  // Top level resolvers
+  Query: queryResolvers,
+  // Type resolvers
   ...typeResolvers
 };
